+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix laat word going after tool call for openai-chat API.
+
 ## 0.63.2
 
 - Fix retrocompatibility with some models not working with openai-chat like deepseek.